Improve zero handling in plot_bipartite.
Fix a bug that occurs in plot_bipartite when the consumer data set has a zero column sum for a resource species (i.e. no interations with that species). Add unit tests to check this revision.

In addition, update reference tables used in unit testing for bipartite_stats to reflect recent updates to the bipartite package.